quantized elu: require observation
Summary:

ELU has a range of [-1, inf]. In the original PR which added
the quantized operator we decided to pass the quantization params
from the input.  However, it makes more sense to require observation
for this op.

This PR changes the API to require observation. Next PRs in this stack
will add the eager and graph mode handling.

Test Plan:

```
python test/test_quantization.py TestQuantizedOps.test_qelu
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: ce9fb77d76009912f6811a80159fef29e4a2fcda
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40100