Add finishAndThrow function to ProcessGroup::Work, and use with Gloo
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40405

This adds a finishAndThrow function that completes the work object,
sets an exception if one is provided by the user, and throws an exception (if
it is already set or passed by the caller). This is now done by grabbing the
lock just once and simplifies the wait functions in ProcessGroupGloo.
ghstack-source-id: 106381735

Differential Revision: [D22174890](https://our.internmc.facebook.com/intern/diff/D22174890/)